move idpmanager to its own package
