--- management/server/idpmanager.go
@@ -1,7 +1,0 @@
-package server
-
-type IDPManager interface {
-	UpdateUserMetadata(userId string, metadata Metadata) error
-}
-
-type Metadata map[string]string
